refactor: divide internalGeneralBody in two components
